pyproject.toml
[tool.poetry]
name = "prediction-market-tools"
version = "0.0.1"
description = ""
authors = ["Zack Drescher <zackryandrescher@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kalshi-python = "^2.0.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
pydantic-settings = "^2.5.2"
cryptography = "^43.0.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
ruff = "^0.6.9"
requests = "^2.32.3"
pandas = "^2.2.3"
mypy = "^1.13.0"
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ['pydantic.mypy']
